Code Refactor
- moved files in to folder and split them up
- and update read me with new install instruction
