+++ README.md
@@ -1,4 +1,5 @@
 # Shady-Pencil
+
 convert blender grease pencil animations to mesh animations to use the shader editor
 
 ## **WARNING**
@@ -6,18 +7,19 @@ convert blender grease pencil animations to mesh animations to use the shader ed
 ** IMPORTANT NEW UPDATE WATCH THE VIDEO IN HOW TO USE IF IT'S NOT WORKING FOR YOU MIGTH HELP **
 
 the code is slow and can be unstable save often so you down loss you're progress
-i have a 4670k and my animation that i created take **25 sec** to run 
+i have a 4670k and my animation that i created take **25 sec** to run
 
 **your one might take longer**
 
 ## DOWNLOAD AND RUN
 
-1. Click the Code dropdown on github page 
-2. Download Zip 
+1. Click the Code dropdown on github page
+2. Download Zip
 3. unzip and run the **Shady_Pencil.py** file in blender
-4. read the instructions
+4. move the **Shady_pencil** forlder into the add on directory of you're blender install should be something like C:\programs\Blender Foundation\Blender 3.3\3.3\scripts\addons in windows For Some Reason **ZIP INSTALLS** don't work
+5. read the instructions
 
-## HOW TO USE 
+## HOW TO USE
 
 go watch this youtube video of a demo sorry for the bad audio ill fix it in the future
 [![Shady Pencil]()](https://youtu.be/7wY5nFMHoSI)
