+++ Shady_pencil/__init__.py
@@ -1,0 +1,27 @@
+bl_info = {
+    "name": "Shady Pencil",
+    "author": "AHK <https://github.com/ahkCodeder>",
+    "version": (1,0),
+    "blender": (3,3,0),
+    "category": "3D View",
+    "location": "View3D",
+    "description": "This turns Grease pencil strokes into Mesh Objects",
+    "warning": "Lines can't be turned to Mesh obj",
+    "doc_url": "https://github.com/ahkCodeder"
+}
+
+import bpy
+from Shady_pencil import operator, panel
+
+def register():
+
+    operator.register()
+    panel.register()
+
+def unregister():
+
+    operator.unregister()
+    panel.unregister()
+
+if __name__ == "__main__":
+    register()
